Extend OCInvocationCallbacks into NativeLong, and short. Add CFIndex, CFRange, NSInteger, NSRange, NSUInteger to match core typedefs and make us 64-bit safe(r). That's safer, not trademark. Thanks to AndyT.

+++ rococoa/src/test/java/org/rococoa/quicktime/QTMovieTest.java
@@ -67,7 +67,7 @@ public class QTMovieTest extends NSTestCase {
         QTMovie movie = QTMovie.movieWithFile_error(file, errorReference);
         assertTrue(movie.id().isNull());
         NSError error = errorReference.getValueAs(NSError.class);
-        assertEquals(-2000, error.code());
+        assertEquals(-2000, error.code().intValue());
     }
 
     public void testAttributeForKey() throws Exception {
